SDC. One more uint8_t, uint32_t mixture fixed.
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/branches/sdc_dev2@4101 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/hal/src/sdc.c
@@ -52,7 +52,7 @@
  *
  * @notapi
  */
-static uint32_t _sdc_get_slice(uint32_t *data, int8_t end, int8_t start) {
+static uint32_t _sdc_get_slice(uint32_t *data, int32_t end, int32_t start) {
   uint32_t word = 0;
   uint32_t mask = 0;
